Bug fix for PFR failing with inf and nan values

+++ src/pfr1d_solver.h
@@ -115,7 +115,7 @@ protected:
     double m_h0;
 
     //! Max solver internal steps 
-    double m_maxSteps;
+    size_t m_maxSteps;
 
     //! Simulation end time
     double m_tStop;
